IMPALA-10441: Skip test_bytes_read_per_column if not on local minicluster
IMPALA-9865 part 2 made some expected outcomes of the above test steps
stricter. Unfortunately these stricter results are only valid when the
tests are run on an HDFS file system in the context of a local
minicluster, breaking the same test on S3 and EC storage.

The patch disables the test step when run outside the context of a local
minicluster HDFS.

Change-Id: If8a179937c9c7c690dd2630549464dbe6aa1b834
Reviewed-on: http://gerrit.cloudera.org:8080/16964
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>