IMPALA-10519: Allow setting of num_reactors for KuduClient
The KuduClient allows setting a number of reactor threads, which are
used for sending rpcs to Kudu such as for scans. The defaults is 4
threads, which may be too few for many use cases, since there is a
single KuduClient per impalad.

This patch adds a flag --kudu_client_num_reactor_threads which allows
users to set the number of threads. The default for the flag is 4,
which reflects the current default within Kudu, to avoid possible
regressions.

Some follow up work could be to do performance experiments and set
the default empirically.

Change-Id: If2ccb9659b9223c9a5de2416b946e6313a3239ff
Reviewed-on: http://gerrit.cloudera.org:8080/17086
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>